+++ Api/tests/UrlShortener.Api.Core.Tests/Base62EncodingScenarios.cs
@@ -11,7 +11,7 @@ public sealed class Base62EncodingScenarios
     [InlineData(61, "z")]
     [InlineData(1000, "G8")]
     [InlineData(987654321, "14q60P")]
-    public void Should_Encode_Number_To_Base62(int number, string expected)
+    public void Should_Encode_Number_To_Base62(long number, string expected)
     {
         number.EncodeToBase62().Should().Be(expected);
     }
